Rename modules in protocol
