+++ core/network/src/lib.rs
@@ -24,19 +24,12 @@
 //!
 
 mod service;
-mod sync;
 #[macro_use]
 mod protocol;
 mod chain;
-mod blocks;
-mod on_demand;
 mod on_demand_layer;
-mod util;
 pub mod config;
-pub mod consensus_gossip;
 pub mod error;
-pub mod message;
-pub mod specialization;
 
 #[cfg(any(test, feature = "test-helpers"))]
 pub mod test;
@@ -46,8 +39,8 @@ pub use service::{
 	NetworkService, NetworkWorker, FetchFuture, TransactionPool, ManageNetwork,
 	NetworkMsg, SyncProvider, ExHashT, ReportHandle,
 };
-pub use protocol::{ProtocolStatus, PeerInfo, Context};
-pub use sync::{Status as SyncStatus, SyncState};
+pub use protocol::{ProtocolStatus, PeerInfo, Context, consensus_gossip, message, specialization};
+pub use protocol::sync::{Status as SyncStatus, SyncState};
 pub use network_libp2p::{
 	identity, multiaddr,
 	ProtocolId, Multiaddr,
@@ -57,7 +50,7 @@ pub use network_libp2p::{
 };
 pub use message::{generic as generic_message, RequestId, Status as StatusMessage};
 pub use error::Error;
-pub use on_demand::AlwaysBadChecker;
+pub use protocol::on_demand::AlwaysBadChecker;
 pub use on_demand_layer::{OnDemand, RemoteResponse};
 #[doc(hidden)]
 pub use runtime_primitives::traits::Block as BlockT;
