+++ core/network/src/protocol/message.rs
