+++ core/network/src/protocol/sync/extra_requests.rs
@@ -23,8 +23,8 @@ use fork_tree::ForkTree;
 use network_libp2p::PeerId;
 use runtime_primitives::Justification;
 use runtime_primitives::traits::{Block as BlockT, NumberFor};
-use crate::message;
-use crate::sync::{Context, PeerSync, PeerSyncState};
+use crate::protocol::message;
+use crate::protocol::sync::{Context, PeerSync, PeerSyncState};
 
 // Time to wait before trying to get the same extra data from the same peer.
 const EXTRA_RETRY_WAIT: Duration = Duration::from_secs(10);
