+++ core/network/src/protocol/util.rs
